embedding type 64

+++ api/types.go
@@ -215,7 +215,7 @@ type EmbeddingRequest struct {
 
 // EmbeddingResponse is the response from [Client.Embeddings].
 type EmbeddingResponse struct {
-	Embedding []float32 `json:"embedding"`
+	Embedding []float64 `json:"embedding"`
 }
 
 // CreateRequest is the request passed to [Client.Create].
